Auto set data in Total of Industrial & Domestic in Disposal Report.
- add parseToDouble extension function
- change datatype to double of 2 Disposal fields in ReportRequest.kt

+++ app/src/main/java/com/example/mpcb/network/request/ReportRequest.kt
@@ -289,10 +289,10 @@ class RoutineReport {
     var disposalDomesticAnyOtherText = ""
     @SerializedName("disposal_industrial_total")
     @Expose
-    var disposalIndustrialTotal: Int? = null
+    var disposalIndustrialTotal: Double? = null
     @SerializedName("disposal_domestic_total")
     @Expose
-    var disposalDomesticTotal: Int? = null
+    var disposalDomesticTotal: Double? = null
     @SerializedName("disposal_observation")
     @Expose
     var disposalObservation = ""
